adding icp workflow
