+++ array_analyzer/extract/image_parser.py
@@ -2,11 +2,11 @@
 
 import os
 from copy import copy
+import cv2 as cv
 import numpy as np
 import re
 import itertools
 import math
-import cv2 as cv
 from types import SimpleNamespace
 from scipy import spatial, stats
 
@@ -65,7 +65,6 @@ def read_to_grey(path_, wellimage_):
 def read_gray_im(im_path):
     """
     Read image from full path to file location.
-
     :param str im_path: Path to image
     :return np.array im: Grayscale image
     """
@@ -178,11 +177,11 @@ def crop_image(arr, cx_, cy_, radius_, border_=200):
     crop the supplied image to include only the well and its spots
 
     :param arr: image
-    :param cx_: float
-    :param cy_: float
-    :param radius_:
-    :param border_:
-    :return:
+    :param float cx_: Center x coordinate
+    :param float cy_: Center y coordinate
+    :param float radius_: Crop radius
+    :param int border_: Margin on each side in pixels
+    :return np.array crop: Cropped image
     """
     cx_ = int(np.rint(cx_))
     cy_ = int(np.rint(cy_))
@@ -395,124 +394,6 @@ def find_fiducials_markers(props_,
     return cent_map
 
 
-def get_spot_coords(im,
-                    min_thresh=0,
-                    max_thresh=255,
-                    min_area=50,
-                    max_area=10000,
-                    min_circularity=0.1,
-                    min_convexity=0.5):
-    """
-    Use OpenCVs simple blob detector (thresholdings and grouping by properties)
-    to detect all dark spots in the image
-
-    :param np.array im: uint8 mage containing spots
-    :param int min_thresh: Minimum threshold
-    :param int max_thresh: Maximum threshold
-    :param int min_area: Minimum spot area in pixels
-    :param int max_area: Maximum spot area in pixels
-    :param float min_circularity: Minimum circularity of spots
-    :param float min_convexity: Minimum convexity of spots
-    :return np.array spot_coords: x, y coordinates of spot centroids (nbr spots x 2)
-    """
-    params = cv.SimpleBlobDetector_Params()
-
-    # Change thresholds
-    params.minThreshold = min_thresh
-    params.maxThreshold = max_thresh
-    # Filter by Area
-    params.filterByArea = True
-    params.minArea = min_area
-    params.maxArea = max_area
-    # Filter by Circularity
-    params.filterByCircularity = True
-    params.minCircularity = min_circularity
-    # Filter by Convexity
-    params.filterByConvexity = True
-    params.minConvexity = min_convexity
-
-    detector = cv.SimpleBlobDetector_create(params)
-
-    # Normalize image
-    im_norm = ((im - im.min()) / (im.max() - im.min()) * 255).astype(np.uint8)
-    # Detect blobs
-    keypoints = detector.detect(im_norm)
-
-    spot_coords = np.zeros((len(keypoints), 2))
-    # Convert to np.arrays
-    for c in range(len(keypoints)):
-        pt = keypoints[c].pt
-        spot_coords[c, 0] = pt[0]
-        spot_coords[c, 1] = pt[1]
-
-    return spot_coords
-
-
-def find_profile_peaks(profile, margin, prominence):
-    # invert because black spots
-    profile = profile.max() - profile
-    max_pos = int(np.where(profile == profile.max())[0])
-    # Make sure max is not due to leaving the center
-    add_margin = 0
-    half_margin = int(margin / 2)
-    if max_pos > len(profile) - half_margin:
-        profile = profile[:-half_margin]
-    elif max_pos < half_margin:
-        profile = profile[half_margin:]
-        add_margin = half_margin
-    profile = gaussian_filter1d(profile, 3)
-    min_prom = profile.max() * prominence
-    peaks, _ = find_peaks(profile, prominence=min_prom, distance=50)
-    if len(peaks) >= 4:
-        spot_dists = peaks[1:] - peaks[:-1]
-    else:
-        spot_dists = None
-    mean_pos = peaks[0] + (peaks[-1] - peaks[0]) / 2 + add_margin
-    return mean_pos, spot_dists
-
-
-def grid_estimation(im,
-                    spot_coords,
-                    margin=50,
-                    prominence=.15):
-    """
-    Based on images intensities and detected spots, make an estimation
-    of grid location so that ICP algorithm is initialized close enough for convergence.
-    TODO: This assumes that you always detect the first peaks
-    this may be unstable so think of other ways to initialize...
-
-    :param np.array im: Grayscale image
-    :param np.array spot_coords: Spot x,y coordinates (nbr spots x 2)
-    :param int margin: Margin for cropping outside all detected spots
-    :param float prominence: Fraction of max intensity to filter out insignificant peaks
-    :return tuple start_point: Min x, y coordinates for initial grid estimate
-    :return float spot_dist: Estimated distance between spots
-    """
-    im_shape = im.shape
-    x_min = int(max(margin, np.min(spot_coords[:, 0]) - margin))
-    x_max = int(min(im_shape[1] - margin, np.max(spot_coords[:, 0]) + margin))
-    y_min = int(max(margin, np.min(spot_coords[:, 1]) - margin))
-    y_max = int(min(im_shape[0] - margin, np.max(spot_coords[:, 1]) + margin))
-    im_roi = im[y_min:y_max, x_min:x_max]
-    # Create intensity profiles along x and y and find peaks
-    profile_x = np.mean(im_roi, axis=0)
-    mean_x, dists_x = find_profile_peaks(profile_x, margin, prominence)
-    profile_y = np.mean(im_roi, axis=1)
-    mean_y, dists_y = find_profile_peaks(profile_y, margin, prominence)
-
-    mean_point = (x_min + mean_x, y_min + mean_y)
-    spot_dist = np.hstack([dists_x, dists_y])
-    # Remove invalid distances
-    spot_dist = spot_dist[np.where(spot_dist != None)]
-    if spot_dist.size == 0:
-        # Failed at estimating spot dist. Return default or error out?
-        spot_dist = 80
-    else:
-        spot_dist = np.median(spot_dist)
-
-    return mean_point, spot_dist
-
-
 # def grid_from_centroids(props_, im, n_rows, n_cols, min_area=100, im_height=2048, im_width=2048):
 def grid_from_centroids(props_, im, n_rows, n_cols, dist_flr=True):
     """
@@ -531,8 +412,6 @@ def grid_from_centroids(props_, im, n_rows, n_cols, dist_flr=True):
     :return: dict
         of format (cent_x, cent_y): prop
     """
-
-
     centroids = np.array([prop.weighted_centroid for prop in props_])
     bbox_area = np.array([prop.bbox_area for prop in props_])
     # calculate mean bbox width for cropping undetected spots
@@ -839,208 +718,3 @@ def compute_od(props_array,bgprops_array):
 
     return od_norm, i_spot, i_bg
 
-
-def create_reference_grid(mean_point,
-                          nbr_grid_rows=6,
-                          nbr_grid_cols=6,
-                          spot_dist=83):
-    """
-    Generate initial spot grid based on image scale and number of spots.
-    :param tuple start_point: (x,y) coordinates of center of grid
-    :param int nbr_grid_rows: Number of spot rows
-    :param int nbr_grid_cols: Number of spot columns
-    :param int spot_dist: Distance between spots
-    :return np.array grid_coords: (x, y) coordinates for reference spots (nbr x 2)
-    """
-    start_x = mean_point[0] - spot_dist * (nbr_grid_cols - 1) / 2
-    start_y = mean_point[1] - spot_dist * (nbr_grid_rows - 1) / 2
-    x_vals = np.linspace(start_x, start_x + (nbr_grid_cols - 1) * spot_dist, nbr_grid_cols)
-    y_vals = np.linspace(start_y, start_y + (nbr_grid_rows - 1) * spot_dist, nbr_grid_rows)
-    grid_x, grid_y = np.meshgrid(x_vals, y_vals)
-    grid_x = grid_x.flatten()
-    grid_y = grid_y.flatten()
-    grid_coords = np.vstack([grid_x.T, grid_y.T]).T
-
-    return grid_coords
-
-
-def icp(source, target, max_iterate=50, matrix_diff=1.):
-    """
-    Iterative closest point. Expects x, y coordinates of source and target in
-    an array with shape: nbr of points x 2
-
-    :param np.array source: Source spot coordinates
-    :param np.array target: Target spot coordinates
-    :param int max_iterate: Maximum number of registration iterations
-    :param float matrix_diff: Sum of absolute differences between transformation
-        matrices after one iteration
-    :return np.array t_matrix: 2D transformation matrix
-    """
-    src = source.copy().astype(np.float32)
-    dst = target.copy().astype(np.float32)
-
-    src = np.expand_dims(src, 0)
-    dst = np.expand_dims(dst, 0)
-
-    # Initialize kNN module
-    knn = cv.ml.KNearest_create()
-    labels = np.array(range(dst.shape[1])).astype(np.float32)
-    knn.train(dst[0], cv.ml.ROW_SAMPLE, labels)
-    # Initialize transformation matrix
-    t_matrix = np.eye(3)
-    t_temp = np.eye(3)
-    t_old = t_matrix
-
-    # Iterate while matrix difference > threshold
-    for i in range(max_iterate):
-
-        # Find closest points
-        ret, results, neighbors, dist = knn.findNearest(src[0], 1)
-        # Outlier removal
-        idxs = np.squeeze(neighbors.astype(np.uint8))
-        dist_max = 2 * np.median(dist)
-        normal_idxs = np.where(dist < dist_max)[0]
-        idxs = idxs[normal_idxs]
-        # Find rigid transform
-        t_iter = cv.estimateRigidTransform(
-            src[0, normal_idxs, :],
-            dst[0, idxs, :],
-            fullAffine=False,
-        )
-        if t_iter is None:
-            print("Optimization failed. Using initial estimate")
-            return np.eye(3)[:2]
-        t_temp[:2] = t_iter
-        src = cv.transform(src, t_iter)
-        t_matrix = np.dot(t_temp, t_matrix)
-        # Estimate diff
-        t_diff = sum(sum(abs(t_matrix[:2] - t_old[:2])))
-        t_old = t_matrix
-        if t_diff < matrix_diff:
-            break
-
-    return t_matrix[:2]
-
-
-def create_uniform_particles(x_range, y_range, hdg_range, N):
-    particles = np.empty((N, 3))
-    particles[:, 0] = np.random.uniform(x_range[0], x_range[1], size=N)
-    particles[:, 1] = np.random.uniform(y_range[0], y_range[1], size=N)
-    particles[:, 2] = np.random.uniform(hdg_range[0], hdg_range[1], size=N)
-    particles[:, 2] %= 2 * np.pi
-    return particles
-
-
-def create_gaussian_particles(x_vars,
-                              y_vars,
-                              scale_vars=(1, .1),
-                              angle_vars=(0, .5),
-                              nbr_particles=100):
-    """
-    Create particles from parameters x, y, scale and angle given mean and std.
-
-    :param x_vars:
-    :param y_vars:
-    :param scale_vars:
-    :param angle_vars:
-    :param nbr_particles:
-    :return:
-    """
-    # x, y, scale, angle
-    particles = np.empty((nbr_particles, 4))
-    particles[:, 0] = x_vars[0] + (np.random.randn(nbr_particles) * y_vars[1])
-    particles[:, 1] = y_vars[0] + (np.random.randn(nbr_particles) * y_vars[1])
-    particles[:, 2] = angle_vars[0] + (np.random.randn(nbr_particles) * angle_vars[1])
-    particles[:, 3] = scale_vars[0] + (np.random.randn(nbr_particles) * scale_vars[1])
-    return particles
-
-
-def particle_filter(fiducial_coords,
-                    spot_coords,
-                    particles,
-                    stds,
-                    max_iter=50,
-                    stop_criteria=.1):
-    """
-    Particle filtering to determine best grid location
-
-    :param fiducial_coords:
-    :param spot_coords:
-    :param particles:
-    :param stds:
-    :param max_iter:
-    :param stop_criteria:
-    :return:
-    """
-
-    # Pretrain spot coords
-    dst = spot_coords.copy().astype(np.float32)
-    knn = cv.ml.KNearest_create()
-    labels = np.array(range(dst.shape[0])).astype(np.float32)
-    knn.train(dst, cv.ml.ROW_SAMPLE, labels)
-
-    nbr_particles = particles.shape[0]
-    weights = np.zeros(nbr_particles)
-
-    # Iterate until min dist doesn't change
-    min_dist_old = 10 ** 6
-    for i in range(max_iter):
-        # Reduce standard deviations a little every iteration
-        stds = stds * 0.95 ** i
-
-        # im_roi = image.copy()
-        # im_roi = cv.cvtColor(im_roi, cv.COLOR_GRAY2RGB)
-
-        for p in range(nbr_particles):
-            particle = particles[p]
-            # Generate transformation matrix
-            t_matrix = cv.getRotationMatrix2D(
-                (particle[0], particle[1]),
-                particle[2],
-                particle[3],
-            )
-            trans_coords = cv.transform(np.array([fiducial_coords]), t_matrix)
-            trans_coords = trans_coords[0].astype(np.float32)
-
-            # for c in range(trans_coords.shape[0]):
-            #     coord = tuple(trans_coords[c, :].astype(np.int))
-            #     cv.circle(im_roi, coord, 2, (0, 0, 255), 10)
-
-            # Find nearest spots
-            ret, results, neighbors, dist = knn.findNearest(trans_coords, 1)
-            weights[p] = np.linalg.norm(dist)
-
-        # plt.imshow(im_roi)
-        # plt.axis('off')
-        # plt.show()
-
-        min_dist = np.min(weights)
-        print(min_dist)
-        # Low distance should correspond to high probability
-        weights = 1 / weights
-        # Make weights sum to 1
-        weights = weights / sum(weights)
-
-        # Importance sampling
-        idxs = np.random.choice(nbr_particles, nbr_particles, p=weights)
-        particles = particles[idxs, :]
-
-        # Distort particles
-        for c in range(4):
-            distort = np.random.randn(nbr_particles)
-            particles[:, c] = particles[:, c] + distort * stds[c]
-
-        # See if min dist is not decreasing anymore
-        if abs(min_dist_old - min_dist) < stop_criteria:
-            break
-        min_dist_old = min_dist
-
-    # Return best particle
-    particle = particles[weights == weights.max(), :][0]
-    # Generate transformation matrix
-    t_matrix = cv.getRotationMatrix2D(
-        (particle[0], particle[1]),
-        particle[2],
-        particle[3],
-    )
-    return t_matrix
